Completed conditionals implementation for jump server deployment

+++ config/jump_server/destroy.sh
@@ -4,11 +4,17 @@ set -euo pipefail
 
 STACK_NAME=jump-server
 
-echo Destroying Stack $STACK_NAME...
+STACK_EXISTS=$(aws cloudformation describe-stack-instances --filters Name=tag:Name,Values="$STACK_NAME" Name=instance-state-code,Values=16 --output text)
+echo "$STACK_EXISTS"
+if ["$STACK_EXISTS" != ""] then
+    echo Destroying Stack $STACK_NAME...
 
-aws cloudformation delete-stack \
+    aws cloudformation delete-stack \
         --stack-name "$STACK_NAME" 
 
-echo
-echo "Jump server with name $STACK_NAME destroyed."
-echo
+    echo
+    echo "Jump server with name $STACK_NAME destroyed."
+    echo
+else
+    echo "Stack $STACK_NAME does not exist."
+    echo
